Big Refactor Domain and DB, Auth JWT

+++ Medifix.Application/Locations/DeleteLocation/DeleteLocationCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediFix.Application.Abstractions.Messaging;
+using MediFix.Domain.Locations;
 using MediFix.SharedKernel.Results;
 
 namespace MediFix.Application.Locations.DeleteLocation;
@@ -9,6 +10,19 @@ internal sealed class DeleteLocationCommandHandler(
 {
     public async Task<Result> Handle(DeleteLocationCommand request, CancellationToken cancellationToken)
     {
+        // TODO: Ensure the location can be deleted
+        if (!LocationCanBeDeleted(request.LocationId))
+        {
+            return Error.Validation(
+                "Location.DeleteNotAllowed",
+                "The location cannot be deleted.");
+        }
+
         return await locationsRepository.DeleteByIdAsync(request.LocationId, cancellationToken);
+    }
+
+    private bool LocationCanBeDeleted(LocationId locationId)
+    {
+        throw new NotImplementedException();
     }
 }
